portico: Revert email field to disabled input with plain text styling.
Due to previous changes of making email field no longer a disabled input,
password managers like Lastpass started taking 'Organization name' as
username. But we still need to avoid displaying email as a disabled input
for better UX.

See https://chat.zulip.org/#narrow/stream/6-frontend/topic/realm.20creation.20form/near/944139.

So, email field in HTML is changed back to disabled input allowing password
managers to consider it as a username field again but is styled so as to
display it as a plain text in browsers rather than as a disabled input,
thus, keeping the UX.